Move some tests into plugin_test_{unix,windows}.go, to deal with different behavior of echo
